Added more safety checks with the DownloadDirectory

+++ BetterMajorasMaskInstaller/Window/Completed.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using File = System.IO.File;
 
 namespace BetterMajorasMaskInstaller.Window
 {
@@ -19,6 +20,7 @@ namespace BetterMajorasMaskInstaller.Window
             InitializeComponent();
         }
         public void Welcome_Closing(object sender, CancelEventArgs e) => QuitButton_Click(this, null);
+        public InstallerComponents InstallerComponents { get; set; }
 
         private void QuitButton_Click(object sender, EventArgs e)
         {
@@ -45,8 +47,38 @@ namespace BetterMajorasMaskInstaller.Window
 
             if(TemporaryFilesCheckBox.Checked)
             {
-                if(Directory.Exists(InstallerSettings.DownloadDirectory))
-                    Directory.Delete(InstallerSettings.DownloadDirectory, true);
+                // delete each file from each component,
+                // then delete the directory if it's empty
+                foreach (InstallerComponent component in InstallerComponents.Components)
+                {
+                    foreach (UrlInfo file in component.Urls)
+                    {
+                        string targetFile = Path.Combine(InstallerSettings.DownloadDirectory,
+                                    file.FileName);
+
+                        if (File.Exists(targetFile))
+                            File.Delete(targetFile);
+                    }
+                }
+
+                // also delete extracted 7za.exe and project64.zip
+                foreach (string file in new string[] { "7za.exe", "Project64.zip" })
+                {
+                    string targetFile = Path.Combine(InstallerSettings.DownloadDirectory, file);
+
+                    if (File.Exists(targetFile))
+                        File.Delete(targetFile);
+                }
+
+                // hmm, the dir can be gone? 
+                if (Directory.Exists(InstallerSettings.DownloadDirectory))
+                {
+                    // make sure directory is empty
+                    if (!Directory.EnumerateFileSystemEntries(InstallerSettings.DownloadDirectory).Any())
+                    {
+                        Directory.Delete(InstallerSettings.DownloadDirectory, true);
+                    }
+                }
             }
 
             Application.Exit();
